+++ packages/components/atoms/f-card/test/specs/accessibility/axe-accessibility.spec.js
@@ -1,0 +1,15 @@
+import { getAccessibilityTestResults } from '../../../../../../../test/utils/axe-helper';
+import CardComponent from '../../../test-utils/component-objects/f-card.component';
+
+describe('Accessibility tests', () => {
+    beforeEach(() => {
+        browser.url('?path=/story/components-atoms--card-component');
+        browser.switchToFrame(0);
+        CardComponent.waitForCard();
+    });
+
+    it('a11y - should test f-card component WCAG compliance', () => {
+        // Act
+        const axeResults = getAccessibilityTestResults('f-card');
+    });
+});