+++ packages/components/molecules/f-mega-modal/jest.config.js
@@ -32,6 +32,10 @@ module.exports = {
         }
     },
 
+    modulePathIgnorePatterns: [
+        './test/specs/accessibility'
+    ],
+
     testURL: 'http://localhost/',
 
     setupFilesAfterEnv: [