+++ packages/components/organisms/f-footer/jest.config.js
@@ -30,7 +30,8 @@ module.exports = {
     },
 
     modulePathIgnorePatterns: [
-        './test/specs/component/'
+        './test/specs/component/',
+        './test/specs/accessibility'
     ],
 
     setupFilesAfterEnv: [