+++ packages/components/molecules/f-searchbox/CHANGELOG.md
@@ -3,6 +3,15 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v4.0.0-beta.14
+------------------------------
+*January 07, 2021*
+
+### Added
+- Loqate core functionality, initial Axios call & display of results.
+- Tests to cover Loqate functionality.
+
+
 v4.0.0-beta.13
 ------------------------------
 *January 05, 2021*